functions/booli/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==12.12.0
azure-identity
azure-keyvault-secrets
azure-storage-file-datalake
azure-storage-queue
pandas
numpy
BeautifulSoup4
requests
pyarrow
datetime
lxml